+ classes inherit form abstract class
