+++ pps-lab01-intellij-basic-example/src/lab01/example/model/SimpleBankAccountAbstract.java
@@ -1,0 +1,49 @@
+package lab01.example.model;
+
+/**
+ * This class represent a particular instance of a BankAccount.
+ * In particular, a Simple Bank Account allows always the deposit
+ * while the withdrawal is allowed only if the balance greater or equal the withdrawal amount
+ */
+public class SimpleBankAccountAbstract implements BankAccount {
+
+    private double balance;
+    private final AccountHolder holder;
+
+    public SimpleBankAccountAbstract(final AccountHolder holder, final double balance) {
+        this.holder = holder;
+        this.balance = balance;
+    }
+
+    @Override
+    public AccountHolder getHolder(){
+        return this.holder;
+    }
+
+    @Override
+    public double getBalance() {
+        return this.balance;
+    }
+
+    @Override
+    public void deposit(final int userID, final double amount) {
+        if (checkUser(userID)) {
+            this.balance += amount;
+        }
+    }
+
+    @Override
+    public void withdraw(final int userID, final double amount) {
+        if (checkUser(userID) && isWithdrawAllowed(amount)) {
+            this.balance -= amount;
+        }
+    }
+
+    private boolean isWithdrawAllowed(final double amount){
+        return this.balance >= amount;
+    }
+
+    private boolean checkUser(final int id) {
+        return this.holder.getId() == id;
+    }
+}
